[FLINK-20663][core] Forbid calling HybridMemorySegment#wrap on unsafe segments.
